+++ proc-macro-definitions/src/part/html_comment.rs
@@ -1,3 +1,4 @@
+use super::ParentParameterParser;
 use crate::{
 	asteracea_ident,
 	storage_context::{ParseContext, ParseWithContext},
@@ -18,6 +19,7 @@ impl ParseWithContext for HtmlComment {
 	fn parse_with_context(
 		input: ParseStream<'_>,
 		_cx: &mut ParseContext,
+		_: &mut dyn ParentParameterParser,
 	) -> syn::Result<Self::Output> {
 		let open_span;
 		let text;