+++ proc-macro-definitions/src/part/mod.rs
@@ -3,18 +3,21 @@ mod box_expression;
 mod bump_format_shorthand;
 mod capture_definition;
 mod component;
+mod content;
 mod defer;
 mod event_binding;
 mod html_comment;
 mod html_definition;
+
+pub use component::{BlockParentParameters, ParentParameterParser};
 
 //TODO: Renamed module and struct to `element_expression` / `ElementExpression`, factor out text expressions and value expressions.
 //TODO: Rust expressions shouldn't automatically be blocks except for ones after `with`.
 
 pub use self::capture_definition::CaptureDefinition;
 use self::{
-	bind::Bind, box_expression::BoxExpression, component::Component, defer::Defer,
-	html_comment::HtmlComment, html_definition::HtmlDefinition,
+	bind::Bind, box_expression::BoxExpression, component::Component, content::Content,
+	defer::Defer, html_comment::HtmlComment, html_definition::HtmlDefinition,
 };
 use crate::{
 	asteracea_ident,
@@ -43,6 +46,7 @@ pub(crate) enum Part<C: Configuration> {
 	Box(BoxExpression<C>),
 	BumpFormat(BumpFormat),
 	Capture(CaptureDefinition<C>),
+	Content(Content),
 	Comment(HtmlComment),
 	Component(Component<C>),
 	Defer(Defer<C>),
@@ -90,6 +94,7 @@ impl<C: Configuration> Part<C> {
 			| Part::Capture(_)
 			| Part::Comment(_)
 			| Part::Component(_)
+			| Part::Content(_)
 			| Part::Defer(_)
 			| Part::RustBlock(_, _)
 			| Part::Html(_)
@@ -136,6 +141,7 @@ impl<C: Configuration> Parse for Part<C> {
 		Self::parse_with_context(
 			input,
 			&mut ParseContext::new_fragment(&Visibility::Inherited, &Generics::default()),
+			&mut BlockParentParameters,
 		)
 		.and_then(|part| {
 			if let Some(part) = part {
@@ -154,9 +160,10 @@ impl<C: Configuration> Part<C> {
 	pub fn parse_required_with_context(
 		input: ParseStream<'_>,
 		cx: &mut ParseContext,
+		parent_parameter_parser: &mut dyn ParentParameterParser,
 	) -> Result<Self> {
 		let span = input.span();
-		Self::parse_with_context(input, cx).and_then(|part| {
+		Self::parse_with_context(input, cx, parent_parameter_parser).and_then(|part| {
 			if let Some(part) = part {
 				Ok(part)
 			} else {
@@ -171,14 +178,36 @@ impl<C: Configuration> Part<C> {
 
 impl<C: Configuration> ParseWithContext for Part<C> {
 	type Output = Option<Self>;
-	fn parse_with_context(input: ParseStream<'_>, cx: &mut ParseContext) -> Result<Self::Output> {
+	fn parse_with_context(
+		input: ParseStream<'_>,
+		cx: &mut ParseContext,
+		parent_parameter_parser: &mut dyn ParentParameterParser,
+	) -> Result<Self::Output> {
 		let lookahead = input.lookahead1();
 		if lookahead.peek(bind::kw::bind) {
-			Some(Part::Bind(Bind::parse_with_context(input, cx)?))
+			Some(Part::Bind(Bind::parse_with_context(
+				input,
+				cx,
+				parent_parameter_parser,
+			)?))
 		} else if lookahead.peek(Token![box]) {
-			Some(Part::Box(BoxExpression::parse_with_context(input, cx)?))
+			Some(Part::Box(BoxExpression::parse_with_context(
+				input,
+				cx,
+				parent_parameter_parser,
+			)?))
+		} else if lookahead.peek(Token![..]) {
+			Some(Part::Content(Content::parse_with_context(
+				input,
+				cx,
+				parent_parameter_parser,
+			)?))
 		} else if lookahead.peek(defer::kw::defer) {
-			Some(Part::Defer(Defer::parse_with_context(input, cx)?))
+			Some(Part::Defer(Defer::parse_with_context(
+				input,
+				cx,
+				parent_parameter_parser,
+			)?))
 		} else if lookahead.peek(LitStr) {
 			Some(Part::Text(input.parse()?))
 		} else if lookahead.peek(Token![<]) {
@@ -187,14 +216,16 @@ impl<C: Configuration> ParseWithContext for Part<C> {
 				input.parse::<Token![<]>().unwrap();
 				input.parse::<TokenTree>()?
 			} {
-				TokenTree::Punct(punct) if punct.as_char() == '!' => {
-					Some(Part::Comment(HtmlComment::parse_with_context(input, cx)?))
-				}
-				TokenTree::Punct(punct) if punct.as_char() == '*' => {
-					Some(Part::Component(Component::parse_with_context(input, cx)?))
-				}
+				TokenTree::Punct(punct) if punct.as_char() == '!' => Some(Part::Comment(
+					HtmlComment::parse_with_context(input, cx, parent_parameter_parser)?,
+				)),
+				TokenTree::Punct(punct) if punct.as_char() == '*' => Some(Part::Component(
+					Component::parse_with_context(input, cx, parent_parameter_parser)?,
+				)),
 				_ => Some(Part::Html(HtmlDefinition::<C>::parse_with_context(
-					input, cx,
+					input,
+					cx,
+					parent_parameter_parser,
 				)?)),
 			}
 		} else if input.peek(Token![if]) {
@@ -217,7 +248,9 @@ impl<C: Configuration> ParseWithContext for Part<C> {
 				InitMode::Dyn(_) => {
 					todo!("`dyn if`")
 				}
-				InitMode::Spread(_) => Part::parse_required_with_context(input, cx)?,
+				InitMode::Spread(_) => {
+					Part::parse_required_with_context(input, cx, &mut BlockParentParameters)?
+				}
 			};
 			let (else_, else_arm) = if let Some(else_) = input.parse().unwrap() {
 				(
@@ -226,7 +259,11 @@ impl<C: Configuration> ParseWithContext for Part<C> {
 						InitMode::Dyn(_) => {
 							todo!("`dyn if else`")
 						}
-						InitMode::Spread(_) => Part::parse_required_with_context(input, cx)?,
+						InitMode::Spread(_) => Part::parse_required_with_context(
+							input,
+							cx,
+							&mut BlockParentParameters,
+						)?,
 					}
 					.pipe(Box::new),
 				)
@@ -234,7 +271,7 @@ impl<C: Configuration> ParseWithContext for Part<C> {
 				(
 					Token![else](if_.span),
 					call2_for_syn::call2_strict(quote_spanned!(if_.span=> []), |input| {
-						Part::parse_required_with_context(input, cx)
+						Part::parse_required_with_context(input, cx, &mut BlockParentParameters)
 					})
 					.debugless_unwrap()
 					.unwrap()
@@ -282,7 +319,11 @@ impl<C: Configuration> ParseWithContext for Part<C> {
 						InitMode::Dyn(_) => {
 							todo!("`dyn match arm`")
 						}
-						InitMode::Spread(_) => Part::parse_required_with_context(input, cx)?,
+						InitMode::Spread(_) => Part::parse_required_with_context(
+							input,
+							cx,
+							&mut BlockParentParameters,
+						)?,
 					}
 					.pipe(Box::new);
 					arms.push((attrs, pats, guard, fat_arrow, part))
@@ -302,7 +343,8 @@ impl<C: Configuration> ParseWithContext for Part<C> {
 			|| lookahead.peek(Token![|])
 		{
 			if C::CAN_CAPTURE {
-				CaptureDefinition::parse_with_context(input, cx)?.map(Part::Capture)
+				CaptureDefinition::parse_with_context(input, cx, parent_parameter_parser)?
+					.map(Part::Capture)
 			} else {
 				return Err(Error::new(
 					lookahead.error().span(),
@@ -312,9 +354,14 @@ impl<C: Configuration> ParseWithContext for Part<C> {
 		} else if lookahead.peek(Bracket) {
 			let content;
 			let bracket = bracketed!(content in input);
+
+			parent_parameter_parser.parse_any(input, cx)?;
+
 			let mut inner_parts = Vec::new();
 			while !content.is_empty() {
-				if let Some(inner_part) = Part::<C>::parse_with_context(&content, cx)? {
+				if let Some(inner_part) =
+					Part::<C>::parse_with_context(&content, cx, &mut BlockParentParameters)?
+				{
 					inner_parts.push(inner_part);
 				}
 			}
@@ -332,7 +379,7 @@ impl<C: Configuration> ParseWithContext for Part<C> {
 				#let with
 				#let block
 			};
-			let part = Part::parse_with_context(input, cx)?.map(Box::new);
+			let part = Part::parse_with_context(input, cx, parent_parameter_parser)?.map(Box::new);
 			Some(Part::With(with, block, part))
 		} else {
 			return Err(Error::new(
@@ -369,7 +416,8 @@ impl<C: Configuration> Part<C> {
 				tokens
 			}
 			Part::Comment(html_comment) => html_comment.part_tokens(),
-			Part::Component(component) => component.part_tokens(),
+			Part::Component(component) => component.part_tokens(cx)?,
+			Part::Content(content) => content.part_tokens(),
 			Part::Defer(defer) => defer.part_tokens(cx)?,
 			Part::Text(lit_str) => {
 				let asteracea = asteracea_ident(lit_str.span());