+++ tests/infer_builder.rs
@@ -1,0 +1,13 @@
+use asteracea::__::{infer_builder, AnonymousContentParentParameters};
+
+#[test]
+fn my_test() {
+	let _: AnonymousContentParentParameters = {
+		let phantom = [];
+		if false {
+			<[_; 0] as IntoIterator>::into_iter(phantom).next().unwrap()
+		} else {
+			infer_builder(phantom).build()
+		}
+	};
+}